refactor(ui): rename use-today to use-time
