+++ src/ui/app/components/usage-card.tsx
@@ -3,7 +3,7 @@ import { useCallback, useMemo } from "react";
 import { Button } from "@/components/ui/button";
 import { ChevronLeft, ChevronRight } from "lucide-react";
 import { DurationText } from "@/components/time/duration-text";
-import { useToday } from "@/hooks/use-today";
+import { useToday } from "@/hooks/use-time";
 import type { Interval } from "@/lib/time";
 import type { Period } from "@/lib/time";
