+++ src/ui/app/hooks/use-trigger-info.tsx
@@ -1,5 +1,5 @@
 import { useMemo } from "react";
-import { usePeriodInterval } from "@/hooks/use-today";
+import { usePeriodInterval } from "@/hooks/use-time";
 import { timeFrameToPeriod, type Target, type TimeFrame } from "@/lib/entities";
 import { useAppDurations } from "@/hooks/use-repo";
 import { useTargetApps } from "@/hooks/use-refresh";
